llm-rag/requirements.txt
# Haystack RAG Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.1

# MySQL connector
pymysql==1.1.0
cryptography==41.0.7

# Haystack Core
haystack-ai==2.0.0
opensearch-haystack==2.0.0

# Embeddings
sentence-transformers==2.2.2

# Document Converters
pypdf==3.17.0
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.1.3

# LangChain Dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# OpenAI (for both services)
openai==1.6.1

# Environment variables
python-dotenv==1.0.0